refactor(web): extract dashboard page helpers
