+++ apps/web/src/app/(dashboard)/cars/makes/[make]/page.tsx
@@ -1,20 +1,18 @@
-import { redis } from "@sgcarstrends/utils";
 import slugify from "@sindresorhus/slugify";
 import { loadSearchParams } from "@web/app/(dashboard)/cars/makes/[make]/search-params";
 import { MakeDetail } from "@web/components/cars/makes";
 import { StructuredData } from "@web/components/structured-data";
-import { LAST_UPDATED_CARS_KEY, SITE_TITLE, SITE_URL } from "@web/config";
+import { fetchMakePageData } from "@web/lib/cars/make-data";
 import {
   checkMakeIfExist,
   getDistinctMakes,
   getMakeDetails,
 } from "@web/lib/cars/queries";
 import { createPageMetadata } from "@web/lib/metadata";
-import type { Car, Make } from "@web/types";
+import { createWebPageStructuredData } from "@web/lib/metadata/structured-data";
 import { getMonthOrLatest } from "@web/utils/months";
 import type { Metadata } from "next";
 import type { SearchParams } from "nuqs/server";
-import type { WebPage, WithContext } from "schema-dts";
 
 interface Props {
   params: Promise<{ make: string }>;
@@ -72,47 +70,18 @@ const CarMakePage = async ({ params }: Props) => {
       .then((data) => data.logo)
       .catch((e) => console.error(e));
 
-  const [makeExists, makeDetails, makesResult, logo, lastUpdated] =
-    await Promise.all([
-      checkMakeIfExist(make),
-      getMakeDetails(make),
-      getDistinctMakes(),
-      getLogo(),
-      redis.get<number>(LAST_UPDATED_CARS_KEY),
-    ]);
-
-  const makeName = makeExists?.make ?? make.toUpperCase();
-  const makes = makesResult.map((m) => m.make);
-
-  // Transform makeDetails to match expected cars format
-  const cars = {
-    make: makeName,
-    total: makeDetails.total,
-    data: makeDetails.data.map(
-      (d): Car => ({
-        month: d.month,
-        make: makeName,
-        fuel_type: d.fuelType as Car["fuel_type"],
-        vehicle_type: d.vehicleType as Car["vehicle_type"],
-        number: d.count,
-      }),
-    ),
-  };
+  const [{ cars, makes, lastUpdated, makeName }, logo] = await Promise.all([
+    fetchMakePageData(make),
+    getLogo(),
+  ]);
 
   const title = `${makeName} Cars Overview: Registration Trends`;
   const description = `${makeName} cars overview. Historical car registration trends and monthly breakdown by fuel and vehicle types in Singapore.`;
-  const structuredData: WithContext<WebPage> = {
-    "@context": "https://schema.org",
-    "@type": "WebPage",
-    name: title,
+  const structuredData = createWebPageStructuredData(
+    title,
     description,
-    url: `${SITE_URL}/cars/makes/${make}`,
-    publisher: {
-      "@type": "Organization",
-      name: SITE_TITLE,
-      url: SITE_URL,
-    },
-  };
+    `/cars/makes/${make}`,
+  );
 
   return (
     <>
