+++ apps/web/src/app/(dashboard)/cars/makes/page.tsx
@@ -6,7 +6,6 @@ import { StructuredData } from "@web/components/structured-data";
 import { LAST_UPDATED_CARS_KEY, SITE_TITLE, SITE_URL } from "@web/config";
 import { getDistinctMakes } from "@web/lib/cars/queries";
 import { createPageMetadata } from "@web/lib/metadata";
-import type { Make } from "@web/types";
 import { fetchMonthsForCars } from "@web/utils/months";
 import type { Metadata } from "next";
 import type { WebPage, WithContext } from "schema-dts";
