+++ apps/web/src/app/(dashboard)/coe/trends/page.tsx
@@ -1,9 +1,4 @@
-import { redis } from "@sgcarstrends/utils";
-import {
-  getDefaultEndDate,
-  getDefaultStartDate,
-  loadSearchParams,
-} from "@web/app/(dashboard)/coe/search-params";
+import { loadSearchParams } from "@web/app/(dashboard)/coe/search-params";
 import { COECategories } from "@web/components/coe/coe-categories";
 import { COEPremiumChart } from "@web/components/coe/premium-chart";
 import { PageHeader } from "@web/components/page-header";
@@ -16,17 +11,12 @@ import {
   CardHeader,
   CardTitle,
 } from "@web/components/ui/card";
-import { LAST_UPDATED_COE_KEY, SITE_TITLE, SITE_URL } from "@web/config";
-import {
-  calculateTrendInsights,
-  groupCOEResultsByBidding,
-} from "@web/lib/coe/calculations";
-import { getCOEMonths, getCOEResultsFiltered } from "@web/lib/coe/queries";
+import { calculateTrendInsights } from "@web/lib/coe/calculations";
+import { fetchCOEPageData } from "@web/lib/coe/page-data";
 import { createPageMetadata } from "@web/lib/metadata";
-import type { COEBiddingResult, COEResult, Month } from "@web/types";
+import { createWebPageStructuredData } from "@web/lib/metadata/structured-data";
 import type { Metadata } from "next";
 import type { SearchParams } from "nuqs/server";
-import type { WebPage, WithContext } from "schema-dts";
 
 interface Props {
   searchParams: Promise<SearchParams>;
@@ -46,34 +36,14 @@ export const generateMetadata = (): Metadata => {
 
 const COETrendsPage = async ({ searchParams }: Props) => {
   const { start, end } = await loadSearchParams(searchParams);
-  const defaultStart = await getDefaultStartDate();
-  const defaultEnd = await getDefaultEndDate();
-  const startDate = start || defaultStart;
-  const endDate = end || defaultEnd;
-
-  const [coeResults, monthsResult, lastUpdated] = await Promise.all([
-    getCOEResultsFiltered(undefined, startDate, endDate),
-    getCOEMonths(),
-    redis.get<number>(LAST_UPDATED_COE_KEY),
-  ]);
-
-  const months = monthsResult.map(({ month }) => month);
-
-  const data = groupCOEResultsByBidding(coeResults);
+  const { months, lastUpdated, data } = await fetchCOEPageData(start, end);
   const trendInsights = calculateTrendInsights(data);
 
-  const structuredData: WithContext<WebPage> = {
-    "@context": "https://schema.org",
-    "@type": "WebPage",
-    name: title,
+  const structuredData = createWebPageStructuredData(
+    title,
     description,
-    url: `${SITE_URL}/coe/trends`,
-    publisher: {
-      "@type": "Organization",
-      name: SITE_TITLE,
-      url: SITE_URL,
-    },
-  };
+    "/coe/trends",
+  );
 
   return (
     <>
